+++ examples/with-next-siwe-iron-session/src/pages/api/verify.ts
@@ -1,7 +1,9 @@
 import { withIronSessionApiRoute } from 'iron-session/next';
 import { NextApiRequest, NextApiResponse } from 'next';
-import { SiweMessage } from 'siwe';
+import { parseSiweMessage, type SiweMessage } from 'viem/siwe';
+
 import { ironOptions } from '../../../lib/iron';
+import { publicClient } from '../../wagmi';
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   const { method } = req;
@@ -9,17 +11,20 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
     case 'POST':
       try {
         const { message, signature } = req.body;
-        const siweMessage = new SiweMessage(message);
-        const { success, error, data } = await siweMessage.verify({
+        const siweMessage = parseSiweMessage(message) as SiweMessage;
+
+        const success = await publicClient.verifyMessage({
+          address: siweMessage.address,
+          message,
           signature,
         });
 
-        if (!success) throw error;
+        if (!success) throw new Error('Invalid signature.');
 
-        if (data.nonce !== req.session.nonce)
+        if (siweMessage.nonce !== req.session.nonce)
           return res.status(422).json({ message: 'Invalid nonce.' });
 
-        req.session.siwe = data;
+        req.session.siwe = siweMessage;
         await req.session.save();
         res.json({ ok: true });
       } catch (_error) {