+++ packages/example/src/wagmi.ts
@@ -1,6 +1,7 @@
 import { type Chain, getDefaultConfig } from '@rainbow-me/rainbowkit';
 import {
   argentWallet,
+  bestWallet,
   bifrostWallet,
   binanceWallet,
   bitgetWallet,
@@ -36,6 +37,7 @@ import {
   omniWallet,
   oneInchWallet,
   oneKeyWallet,
+  paraSwapWallet,
   phantomWallet,
   rabbyWallet,
   rainbowWallet,
@@ -53,10 +55,12 @@ import {
   uniswapWallet,
   valoraWallet,
   walletConnectWallet,
+  wigwamWallet,
   xdefiWallet,
   zealWallet,
   zerionWallet,
 } from '@rainbow-me/rainbowkit/wallets';
+import { publicActions } from 'viem';
 import {
   arbitrum,
   arbitrumSepolia,
@@ -183,6 +187,7 @@ export const config = getDefaultConfig({
       groupName: 'Other',
       wallets: [
         argentWallet,
+        bestWallet,
         bifrostWallet,
         binanceWallet,
         bitgetWallet,
@@ -216,6 +221,7 @@ export const config = getDefaultConfig({
         omniWallet,
         oneInchWallet,
         oneKeyWallet,
+        paraSwapWallet,
         phantomWallet,
         rabbyWallet,
         ramperWallet,
@@ -230,6 +236,7 @@ export const config = getDefaultConfig({
         trustWallet,
         uniswapWallet,
         valoraWallet,
+        wigwamWallet,
         xdefiWallet,
         zealWallet,
         zerionWallet,
@@ -238,3 +245,5 @@ export const config = getDefaultConfig({
   ],
   ssr: true,
 });
+
+export const publicClient = config.getClient().extend(publicActions);