+++ packages/rainbowkit/build.js
@@ -23,12 +23,6 @@ const getAllEntryPoints = async (rootPath) =>
     );
 
 const baseBuildConfig = (onEnd) => {
-  const rainbowProviderApiKey = process.env.RAINBOW_PROVIDER_API_KEY;
-
-  if (!rainbowProviderApiKey) {
-    throw new Error('missing RAINBOW_PROVIDER_API_KEY env variable');
-  }
-
   return {
     banner: {
       js: '"use client";', // Required for Next 13 App Router
@@ -42,14 +36,6 @@ const baseBuildConfig = (onEnd) => {
       '.json': 'text',
     },
     plugins: [
-      replace({
-        include:
-          /src\/components\/RainbowKitProvider\/useFingerprint.ts$|src\/core\/network\/enhancedProvider.ts$/,
-        values: {
-          __buildVersion: process.env.npm_package_version,
-          __rainbowProviderApiKey: rainbowProviderApiKey,
-        },
-      }),
       vanillaExtractPlugin({
         identifiers: isCssMinified ? 'short' : 'debug',
         processCss: async (css) => {
@@ -82,8 +68,34 @@ const baseBuildConfig = (onEnd) => {
   };
 };
 
+const mainBuildConfig = (onEnd) => {
+  const rainbowProviderApiKey = process.env.RAINBOW_PROVIDER_API_KEY;
+
+  if (!rainbowProviderApiKey) {
+    console.warn(
+      'missing RAINBOW_PROVIDER_API_KEY env variable, disabling enhanced provider',
+    );
+  }
+
+  const buildConfig = baseBuildConfig(onEnd);
+  return {
+    ...buildConfig,
+    plugins: [
+      replace({
+        include:
+          /src\/components\/RainbowKitProvider\/useFingerprint.ts$|src\/core\/network\/enhancedProvider.ts$/,
+        values: {
+          __buildVersion: process.env.npm_package_version,
+          __rainbowProviderApiKey: rainbowProviderApiKey,
+        },
+      }),
+      ...buildConfig.plugins,
+    ],
+  };
+};
+
 const mainBuildOptions = {
-  ...baseBuildConfig((result) => {
+  ...mainBuildConfig((result) => {
     if (result.errors.length) {
       console.error('❌ main build failed:', result.errors);
     } else console.log('✅ main build succeeded');