+++ site/next-sitemap.config.js
@@ -7,6 +7,8 @@ module.exports = {
     '/ds',
     '/ar',
     '/ar/*',
+    '/de',
+    '/de/*',
     '/es-419',
     '/es-419/*',
     '/fr',
@@ -29,6 +31,8 @@ module.exports = {
     '/tr/*',
     '/ua',
     '/ua/*',
+    '/vi',
+    '/vi/*',
     '/zh-CN',
     '/zh-CN/*',
     '/zh-HK',
@@ -44,6 +48,10 @@ module.exports = {
     {
       href: 'https://rainbowkit.com/ar/',
       hreflang: 'ar',
+    },
+    {
+      href: 'https://rainbowkit.com/de/',
+      hreflang: 'de',
     },
     {
       href: 'https://rainbowkit.com/es-419/',
@@ -88,6 +96,10 @@ module.exports = {
     {
       href: 'https://rainbowkit.com/ua/',
       hreflang: 'ua',
+    },
+    {
+      href: 'https://rainbowkit.com/vi/',
+      hreflang: 'vi',
     },
     {
       href: 'https://rainbowkit.com/zh-CN/',