+++ packages/example/CHANGELOG.md
@@ -1,5 +1,27 @@
 # example
 
+## 0.0.107
+
+### Patch Changes
+
+- dc31926: Upgraded `wagmi` to `^2.12.17`
+- Updated dependencies [f02bced]
+- Updated dependencies [f02bced]
+  - @rainbow-me/rainbowkit@2.2.0
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.5.0
+  - @rainbow-me/rainbow-button@0.2.17
+
+## 0.0.106
+
+### Patch Changes
+
+- Updated dependencies [f93cd0e]
+- Updated dependencies [4014d80]
+- Updated dependencies [6393498]
+  - @rainbow-me/rainbowkit@2.1.7
+  - @rainbow-me/rainbow-button@0.2.16
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.4.1
+
 ## 0.0.105
 
 ### Patch Changes