+++ packages/rainbow-button/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @rainbow-me/rainbow-button
 
+## 0.2.17
+
+### Patch Changes
+
+- Updated dependencies [f02bced]
+  - @rainbow-me/rainbowkit@2.2.0
+
+## 0.2.16
+
+### Patch Changes
+
+- Updated dependencies [f93cd0e]
+- Updated dependencies [4014d80]
+- Updated dependencies [6393498]
+  - @rainbow-me/rainbowkit@2.1.7
+
 ## 0.2.15
 
 ### Patch Changes